requirements.txt
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
pyperclip==1.8.2
pystray==0.19.4
flask==2.2.3
Pillow==9.3.0